Corrigindo Services e Controllers para usar Dtos e mappers

+++ src/main/java/org/example/hospitalapi/service/impl/DoctorServiceImpl.java
@@ -1,6 +1,11 @@
 package org.example.hospitalapi.service.impl;
 
+import org.example.hospitalapi.dtos.CreateDoctorDto;
+import org.example.hospitalapi.dtos.DoctorDto;
+import org.example.hospitalapi.dtos.UpdateDoctorDto;
 import org.example.hospitalapi.entity.Doctor;
+import org.example.hospitalapi.exceptions.ResourceNotFoundException;
+import org.example.hospitalapi.mapper.DoctorMapper;
 import org.example.hospitalapi.repository.DoctorRepository;
 import org.example.hospitalapi.service.DoctorService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,7 +13,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.Optional;
 
 @Service
 public class DoctorServiceImpl implements DoctorService {
@@ -16,40 +20,47 @@ public class DoctorServiceImpl implements DoctorService {
     @Autowired
     private DoctorRepository doctorRepository;
 
+    @Autowired
+    private DoctorMapper doctorMapper;
+
     @Override
-    public Page<Doctor> getDoctors(Pageable pageable) {
-        return doctorRepository.findAll(pageable);
+    public Page<DoctorDto> getDoctors(Pageable pageable) {
+        return doctorRepository.findAll(pageable).map(doctorMapper::toDto);
     }
 
     @Override
-    public Doctor createDoctor(Doctor doctor) {
-        return doctorRepository.save(doctor);
+    public DoctorDto createDoctor(CreateDoctorDto createDoctorDto) {
+        Doctor doctor = doctorMapper.toModel(createDoctorDto);
+        return doctorMapper.toDto(doctorRepository.save(doctor));
     }
 
     @Override
-    public Doctor getDoctorById(Long id) {
-        Optional<Doctor> doctor = doctorRepository.findById(id);
-        return doctor.orElse(null);
+    public DoctorDto getDoctorById(Long id) {
+        Doctor doctor = doctorRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Doctor not found with id: " + id));
+        return doctorMapper.toDto(doctor);
     }
 
     @Override
-    public Doctor updateDoctor(Long id, Doctor doctorDetails) {
-        Optional<Doctor> optionalDoctor = doctorRepository.findById(id);
-        if (optionalDoctor.isPresent()) {
-            Doctor doctor = optionalDoctor.get();
-            doctor.setFirstName(doctorDetails.getFirstName());
-            doctor.setLastName(doctorDetails.getLastName());
-            doctor.setRegistration(doctorDetails.getRegistration());
-            doctor.setEmail(doctorDetails.getEmail());
-            doctor.setGender(doctorDetails.getGender());
-            return doctorRepository.save(doctor);
-        } else {
-            throw new IllegalArgumentException("Doctor not found with id: " + id);
-        }
+    public DoctorDto updateDoctor(Long id, UpdateDoctorDto dto) {
+        Doctor doctor = doctorRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Doctor not found with id: " + id));
+
+        if (dto.firstName() != null) doctor.setFirstName(dto.firstName());
+        if (dto.lastName() != null) doctor.setLastName(dto.lastName());
+        if (dto.email() != null) doctor.setEmail(dto.email());
+        if (dto.gender() != null) doctor.setGender(dto.gender());
+
+        doctor = doctorRepository.save(doctor);
+
+        return doctorMapper.toDto(doctor);
     }
 
     @Override
     public void deleteDoctor(Long id) {
+        if (!doctorRepository.existsById(id)) {
+            throw new IllegalArgumentException("Doctor not found with id: " + id);
+        }
         doctorRepository.deleteById(id);
     }
 }
