+++ src/main/java/org/example/hospitalapi/service/impl/PatientServiceImpl.java
@@ -1,6 +1,11 @@
 package org.example.hospitalapi.service.impl;
 
+import org.example.hospitalapi.dtos.CreatePatientDto;
+import org.example.hospitalapi.dtos.UpdatePatientDto;
+import org.example.hospitalapi.dtos.PatientDto;
 import org.example.hospitalapi.entity.Patient;
+import org.example.hospitalapi.exceptions.ResourceNotFoundException;
+import org.example.hospitalapi.mapper.PatientMapper;
 import org.example.hospitalapi.repository.PatientRepository;
 import org.example.hospitalapi.service.PatientService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,7 +13,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.Optional;
 
 @Service
 public class PatientServiceImpl implements PatientService {
@@ -16,38 +20,47 @@ public class PatientServiceImpl implements PatientService {
     @Autowired
     private PatientRepository patientRepository;
 
+    @Autowired
+    private PatientMapper patientMapper;
+
     @Override
-    public Page<Patient> getPatients(Pageable pageable) {
-        return patientRepository.findAll(pageable);
+    public Page<PatientDto> getPatients(Pageable pageable) {
+        return patientRepository.findAll(pageable).map(patientMapper::toDto);
     }
 
     @Override
-    public Patient createPatient(Patient patient) {
-        return patientRepository.save(patient);
+    public PatientDto createPatient(CreatePatientDto createPatientDto) {
+        Patient patient = patientMapper.toModel(createPatientDto);
+        return patientMapper.toDto(patientRepository.save(patient));
     }
 
     @Override
-    public Patient getPatientById(Long id) {
-        Optional<Patient> patient = patientRepository.findById(id);
-        return patient.orElse(null);
+    public PatientDto getPatientById(Long id) {
+        Patient patient = patientRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Patient not found with id: " + id));
+        return patientMapper.toDto(patient);
     }
 
     @Override
-    public Patient updatePatient(Long id, Patient patientDetails) {
-        Optional<Patient> optionalPatient = patientRepository.findById(id);
-        if (optionalPatient.isPresent()) {
-            Patient patient = optionalPatient.get();
-            patient.setFirstName(patientDetails.getFirstName());
-            patient.setEmail(patientDetails.getEmail());
-            patient.setGender(patientDetails.getGender());
-            return patientRepository.save(patient);
-        } else {
-            throw new IllegalArgumentException("Patient not found with id: " + id);
-        }
+    public PatientDto updatePatient(Long id, UpdatePatientDto dto) {
+        Patient patient = patientRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Patient not found with id: " + id));
+
+        if (dto.firstName() != null) patient.setFirstName(dto.firstName());
+        if (dto.lastName() != null) patient.setLastName(dto.lastName());
+        if (dto.email() != null) patient.setEmail(dto.email());
+        if (dto.gender() != null) patient.setGender(dto.gender());
+
+        patient = patientRepository.save(patient);
+
+        return patientMapper.toDto(patient);
     }
 
     @Override
     public void deletePatient(Long id) {
+        if (!patientRepository.existsById(id)) {
+            throw new IllegalArgumentException("Patient not found with id: " + id);
+        }
         patientRepository.deleteById(id);
     }
 }
